Schema and type files were reorganized
