+++ src/schemas/contact-us.schema.ts
@@ -1,4 +1,4 @@
-import { z } from "zod/v4";
+import { z } from 'zod/v4';
 
 export const contactUsFormSchema = z.object({
     email: z.email(),
