+++ src/schemas/index.ts
@@ -1,8 +1,6 @@
-export * from './sign-in';
-export * from './sign-up';
-export * from './add-appointment';
-export * from './profile-picture';
-export * from './user-detail';
-export * from './contact-us';
-export * from './user-filter';
-export * from './feedback-filter';
+export * from './auth.schema';
+export * from './appointment.schema';
+export * from './profile.schema';
+export * from './contact-us.schema';
+export * from './user-filter.schema';
+export * from './feedback-filter.schema';
