+++ src/schemas/profile.schema.ts
@@ -1,5 +1,14 @@
-import { z } from "zod/v4";
+import { z } from 'zod/v4';
+import { ACCEPTED_IMAGE_TYPES, MAX_FILE_SIZE } from '@/constants';
 import languages from '@/languages.json';
+
+export const profilePictureSchema = z
+    .instanceof(File)
+    .refine((file) => file.size <= MAX_FILE_SIZE, 'Max image size is 5 MB')
+    .refine(
+        (file) => ACCEPTED_IMAGE_TYPES.includes(file.type),
+        `Only supported ${ACCEPTED_IMAGE_TYPES.map((type) => type.split('/')[1] || '').join(', ')}`,
+    );
 
 export const userDetailSchema = z
     .object({
