--- src/schemas/sign-in.ts
@@ -1,8 +1,0 @@
-import { z } from "zod/v4";
-
-export const signInSchema = z.object({
-    email: z.string().email('Enter valid email'),
-    password: z
-        .string()
-        .min(8, 'Password must contain at least 8 character(s)'),
-});
