+++ src/schemas/user-filter.schema.ts
@@ -1,7 +1,6 @@
-import { z } from "zod/v4";
+import { z } from 'zod/v4';
 
 export const userFilterSchema = z.object({
     query: z.string().transform((val) => val.toLowerCase()),
-    status: z
-        .enum(['WAITING', 'APPROVED', 'DECLINED', 'ALL'])
+    status: z.enum(['WAITING', 'APPROVED', 'DECLINED', 'ALL']),
 });
