+++ src/types/form.type.ts
@@ -1,4 +1,4 @@
-import { z } from "zod/v4";
+import { z } from 'zod/v4';
 import {
     addAppointmentClientSchema,
     addAppointmentServerSchema,
