+++ src/types/index.ts
@@ -1,6 +1,6 @@
-export * from './sign-out-state.type';
+export * from './sign-out.type';
 export * from './user-without-password';
-export * from './result-with-error.type';
-export * from './form-state-with-errors.type';
-export * from './form.types';
+export * from './result.type';
+export * from './form-state.type';
+export * from './form.type';
 export * from './get-appointments.type';
