+++ src/types/result.type.ts
