--- src/types/sign-out-state.type.ts
@@ -1,3 +1,0 @@
-export type SignOutState = {
-    logout: boolean;
-};
