+++ src/types/sign-out.type.ts
@@ -1,0 +1,3 @@
+export type SignOut = {
+    logout: boolean;
+};
